// src/entities/category/model/types.ts
export type Audience = "girl" | "boy" | "mom" | "accessory" | "gift";

export type ClothingType =
  | "top"
  | "setup"
  | "bottom"
  | "dress"
  | "homewear"
  | "swimwear";

export type CategorySlug =
  | "girl-top"
  | "girl-setup"
  | "girl-bottom"
  | "girl-dress"
  | "girl-homewear"
  | "girl-swimwear"
  | "boy-top"
  | "boy-setup"
  | "boy-bottom"
  | "boy-homewear"
  | "boy-swimwear"
  | "mom"
  | "accessory"
  | "gift";